# docs/swagger.yaml
openapi: 3.0.8
info:
  title: "📚 Golang Book Management API"
  version: 1.0.0
servers:
  - url: "http://{host}/api/"
    variables:
      host:
        enum:
          - localhost:8080
          - localhost:9000
          - mysite.com
        default: localhost:9000
...
paths:
...
  /categories:
    post:
      summary: Create a new category
      tags:
        - Categories
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Category"
      responses:
        201:
          description: Category created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                properties:
                  data:
                    $ref: "#/components/schemas/Category"
    get:
      summary: Get all categories
      tags:
        - Categories
      responses:
        200:
          description: List of categories
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
...
  /books:
    post:
      summary: Create a new book
      tags:
        - Books
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
      responses:
        201:
          description: Book created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                properties:
                  data:
                    $ref: "#/components/schemas/Book"
    get:
      summary: Get all books
      tags:
        - Books
      parameters:
        - in: query
          name: title
          schema:
            type: string
          description: Filter by book title
        - in: query
          name: category_id
          schema:
            type: integer
          description: Filter by category ID
      responses:
        200:
          description: List of books
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/BaseResponse"
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Book"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Category:
      allOf:
        - $ref: "#/components/schemas/BaseProps"
      properties:
        name:
          type: string
    Book:
      allOf:
        - $ref: "#/components/schemas/BaseProps"
      properties:
        title:
          type: string
        description:
          type: string
        image_url:
          type: string
          format: "url"
        relese_year:
          type: integer
          minimum: 1980
          maximum: 2024
        price:
          type: integer
        total_page:
          type: integer
        thickness:
          type: string
          readOnly: true
        category_id:
          type: integer
    BaseProps:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        created_by:
          type: string
          nullable: true
          readOnly: true
        modified_at:
          type: string
          nullable: true
          format: date-time
          readOnly: true
        modified_by:
          type: string
          nullable: true
          format: date-time
          readOnly: true
    BaseResponse:
      type: object
      properties:
        message:
          type: string